pyproject.toml
[tool.poetry]
name = "yellow-snake"
version = "0.1.0"
description = ""
authors = ["forumgamers <forumgamersindo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.63.0"
grpcio-tools = "^1.63.0"
imagekitio = "^4.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
